+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 ## [2.8.0] - 2020/04/30
 
-* Store.waitCondition returns a future which will complete when the given condition is true.
+* Store.waitCondition() returns a future which will complete when the given condition is true.
 The condition can access the state. You may also provide a timeoutInSeconds.
    
-* StoreTester.waitCondition now accepts a parameter called testImmediately.
+* StoreTester.waitCondition() now accepts a parameter called testImmediately.
 When testImmediately is true (now the default), it will test the condition immediately when 
 the method is called. If the condition is true, the method will return immediately,
 without waiting for any actions to be dispatched.